java-pg: Better workaround for Big Sur JVM detection
Trac: https://trac.macports.org/ticket/61445

With this workaround, the check for ${java.version} works as
expected for values like 1.7, 1.8+ or 9 in Big Sur. If the "+"
selector is used and there are more than one candidate, the
highest JVM version wins.

"+" and "*" can only be used on the major version, e.g. 1.8+ or 11*.
Things like 1.8.56* are not supported, as they are not used in a
single port at the moment.